requirements.txt
flask
pyserial
libsass
flask-sqlalchemy
pythonnet
flask-cors
